// src/views/discover/c-views/djradio/store/type.ts
export type djCatelistType = Array<{
  pic56x56Id: number
  pic96x96Id: number
  picPCWhite: number
  picPCBlack: number
  picWeb: number
  picIPad: number
  pic84x84Id: number
  pic56x56IdStr: string
  pic56x56Url: string
  pic96x96IdStr: string
  pic96x96Url: string
  pic84x84IdUrl: string
  picPCWhiteStr: string
  picPCWhiteUrl: string
  picPCBlackStr: string
  picPCBlackUrl: string
  picWebStr: string
  picWebUrl: string
  picMacId: string
  picMacUrl: string
  picUWPId: string
  picUWPUrl: string
  picIPadStr: string
  picIPadUrl: string
  name: string
  id: number
}>
// ...
export type djRecommendByTypeType = Array<{
  dj: {
    defaultAvatar: boolean
    province: number
    authStatus: number
    followed: boolean
    avatarUrl: string
    accountStatus: number
    gender: number
    city: number
    birthday: number
    userId: number
    userType: number
    nickname: string
    signature: string
    description: string
    detailDescription: string
    avatarImgId: number
    backgroundImgId: number
    backgroundUrl: string
    authority: number
    mutual: boolean
    expertTags: any
    experts: any
    djStatus: number
    vipType: number
    remarkName: any
    authenticationTypes: number
    avatarDetail: any
    avatarImgIdStr: string
    backgroundImgIdStr: string
    anchor: boolean
    avatarImgId_str: string
  }
  category: string
  secondCategory: string
  buyed: boolean
  price: number
  originalPrice: number
  discountPrice: any
  purchaseCount: number
  lastProgramName: string
  videos: any
  finished: boolean
  underShelf: boolean
  liveInfo: any
  playCount: number
  privacy: boolean
  icon: any
  manualTagsDTO: any
  descPicList?: Array<{
    type: number
    id: number
    content: string
    height: any
    width: any
    timeStamp: any
    nestedData: {
      textList?: Array<{
        text: string
        attributes: any
      }>
      attributes: any
    }
  }>
  replaceRadioId: number
  replaceRadio: any
  radioFeeType: number
  picId: number
  categoryId: number
  taskId: number
  shortName: any
  intervenePicUrl: string
  programCount: number
  subCount: number
  feeScope: number
  lastProgramId: number
  picUrl: string
  lastProgramCreateTime: number
  intervenePicId: number
  dynamic: boolean
  desc: string
  createTime: number
  name: string
  id: number
  rcmdtext: string
  subed: boolean
  lastUpdateProgramName: string
}>
export type djRadioListType = {
  djRadios: Array<{
    dj: {
      defaultAvatar: boolean
      province: number
      authStatus: number
      followed: boolean
      avatarUrl: string
      accountStatus: number
      gender: number
      city: number
      birthday: number
      userId: number
      userType: number
      nickname: string
      signature: string
      description: string
      detailDescription: string
      avatarImgId: number
      backgroundImgId: number
      backgroundUrl: string
      authority: number
      mutual: boolean
      expertTags?: Array<string>
      experts?: {
        '1': string
      }
      djStatus: number
      vipType: number
      remarkName: any
      authenticationTypes: number
      avatarDetail?: {
        userType: number
        identityLevel: number
        identityIconUrl: string
      }
      avatarImgIdStr: string
      backgroundImgIdStr: string
      anchor: boolean
      avatarImgId_str?: string
    }
    category: string
    secondCategory: string
    buyed: boolean
    price: number
    originalPrice: number
    discountPrice: any
    purchaseCount: number
    lastProgramName: string
    videos: any
    finished: boolean
    underShelf: boolean
    liveInfo: any
    playCount: number
    privacy: boolean
    icon: any
    manualTagsDTO: any
    descPicList?: Array<{
      type: number
      id: number
      content: string
      height: any
      width: any
      timeStamp: any
      nestedData?: {
        textList: Array<{
          text: string
          attributes: any
        }>
        attributes: any
      }
    }>
    replaceRadioId: number
    replaceRadio: any
    shortName: any
    picId: number
    categoryId: number
    taskId: number
    programCount: number
    subCount: number
    feeScope: number
    lastProgramId: number
    picUrl: string
    lastProgramCreateTime: number
    radioFeeType: number
    intervenePicUrl: string
    intervenePicId: number
    dynamic: boolean
    name: string
    id: number
    desc: string
    createTime: number
    rcmdtext: string
    subed: boolean
  }>
  hasMore: boolean
  count: number
  code: number
}

export type IDjradioState = {
  djCatelist: djCatelistType
  djRecommend: djRecommendType
  djRanking: djRankingType
  djRecommendByType: djRecommendByTypeType
  djRadioList: djRadioListType
}